intelligence/osint-tools/domain-intelligence/requirements.txt
# Domain Intelligence System Requirements

# Core Dependencies
requests>=2.28.0
python-whois>=0.8.0
dnspython>=2.3.0
beautifulsoup4>=4.11.0

# SSL/TLS Analysis
pyOpenSSL>=23.0.0
cryptography>=39.0.0

# API Integrations
shodan>=1.28.0

# HTML Parsing
lxml>=4.9.0

# Optional but Recommended
# For better performance and features
urllib3>=1.26.0
certifi>=2022.0.0
charset-normalizer>=3.0.0
idna>=3.4
soupsieve>=2.4
